Xcode 5.0.1 (5A2053)

+++ UIKit.framework/iPhoneOS.platform/UIWebSelectMultiplePicker.h
@@ -18,7 +18,7 @@ __attribute__((visibility("hidden")))
     DOMHTMLSelectElement *_selectionNode;
     NSArray *_cachedItems;
     id <UIWebSelectedItemPrivate> _singleSelectionItem;
-    unsigned int _singleSelectionIndex;
+    unsigned long long _singleSelectionIndex;
     double _fontSize;
     double _maximumTextWidth;
     long long _textAlignment;
@@ -38,7 +38,7 @@ __attribute__((visibility("hidden")))
 - (void)controlBeginEditing;
 - (id)controlView;
 - (void)dealloc;
-- (id)initWithDOMHTMLSelectElement:(id)arg1 cachedItems:(id)arg2 singleSelectionItem:(id)arg3 singleSelectionIndex:(unsigned int)arg4 multipleSelection:(_Bool)arg5;
+- (id)initWithDOMHTMLSelectElement:(id)arg1 cachedItems:(id)arg2 singleSelectionItem:(id)arg3 singleSelectionIndex:(unsigned long long)arg4 multipleSelection:(_Bool)arg5;
 
 @end
